core/views/rest.py
from rest_framework import viewsets

from core import models
from core import serializers


class AssetViewSet(viewsets.ModelViewSet):
    queryset = models.Asset.objects.all()
    serializer_class = serializers.AssetSerializer
