requirements.txt
django==2.0.5
django-bootstrap4==0.0.6
djangorestframework==3.8.2